requirements-dev.txt
-r requirements.txt

# Test / development dependencies
pytest==9.0.2
pytest-flask==1.3.0
